# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Mistri APIs",
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/customer/register": {
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Register customer",
                "operationId": "register",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Customer Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Customer Email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "Customer Password",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password_confirmation",
                        "in": "formData",
                        "description": "Customer Password Confirmation",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "con_number",
                        "in": "formData",
                        "description": "Customer Contact Number",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/customer/login": {
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Login customer",
                "operationId": "login",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Customer Email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "Customer Password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/customer/logout": {
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Logout customer",
                "operationId": "logout",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Auth Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/customer/recover": {
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Recover Customer Password",
                "operationId": "recover",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Customer Email",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/customer/verify-email": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Verify Customer Email",
                "operationId": "verifyEmail",
                "consumes": [
                    "application/xml",
                    "application/json"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "verification_code",
                        "in": "query",
                        "description": "Verification Code",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
        "/api/workshop/recover": {
            "post": {
                "tags": [
                    "Workshops"
                ],
                "summary": "Recover Workshop Password",
                "operationId": "recover",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Workshop Email",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/workshop/verifyEmail": {
            "get": {
                "tags": [
                    "Workshops"
                ],
                "summary": "Verify Workshop Email",
                "operationId": "verifyEmail",
                "consumes": [
                    "application/xml",
                    "application/json"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "verification_code",
                        "in": "query",
                        "description": "Verification Code",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
    },
...
}